# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Roomie API - Chore rotation and expense splitting for shared households
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  householdId: number | null;
  isAdmin: boolean;
}

export interface HouseholdInput {
  /** @minLength 1 */
  name: string;
}

export interface JoinHouseholdInput {
  inviteCode: string;
}

export interface Household {
  id: number;
  name: string;
  inviteCode: string;
  adminUserId: number;
}

export interface Member {
  id: number;
  name: string;
  email: string;
  joinedAt: string;
}

export type ChoreInputFrequency = typeof ChoreInputFrequency[keyof typeof ChoreInputFrequency];


export const ChoreInputFrequency = {
  daily: 'daily',
  weekly: 'weekly',
  custom: 'custom',
} as const;

export interface ChoreInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  frequency: ChoreInputFrequency;
  /**
     * Comma-separated weekday numbers for custom frequency (e.g. "1,3,5")
     * @nullable
     */
  customDays?: string | null;
  /** Ordered list of user IDs */
  rotationOrder: number[];
}

export type ChoreUpdateFrequency = typeof ChoreUpdateFrequency[keyof typeof ChoreUpdateFrequency];


export const ChoreUpdateFrequency = {
  daily: 'daily',
  weekly: 'weekly',
  custom: 'custom',
} as const;

export interface ChoreUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  frequency?: ChoreUpdateFrequency;
  /** @nullable */
  customDays?: string | null;
}

export interface RotationInput {
  rotationOrder: number[];
}

export interface Chore {
  id: number;
  householdId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  frequency: string;
  /** @nullable */
  customDays?: string | null;
  rotationOrder: number[];
}

export type AssignmentStatus = typeof AssignmentStatus[keyof typeof AssignmentStatus];


export const AssignmentStatus = {
  pending: 'pending',
  done: 'done',
} as const;

export interface Assignment {
  id: number;
  choreId: number;
  choreName: string;
  userId: number;
  userName: string;
  date: string;
  status: AssignmentStatus;
  /** @nullable */
  confirmedAt?: string | null;
}

export interface ExpenseInput {
  amount: number;
  /** @minLength 1 */
  description: string;
  /** User ID who paid */
  paidBy: number;
  /** ISO date string, defaults to today */
  date?: string;
  /** User IDs to split among (defaults to all members) */
  splitAmong: number[];
}

export interface ExpenseSplit {
  userId: number;
  userName: string;
  shareAmount: number;
}

export interface Expense {
  id: number;
  householdId: number;
  paidBy: number;
  paidByName: string;
  amount: number;
  description: string;
  date: string;
  splits: ExpenseSplit[];
}

export interface MemberBalance {
  userId: number;
  userName: string;
  /** Positive means they're owed money; negative means they owe money */
  balance: number;
}

export interface SettlementTransaction {
  fromUserId: number;
  fromUserName: string;
  toUserId: number;
  toUserName: string;
  amount: number;
}