# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Roomie API - Chore rotation and expense splitting for shared households
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /chores:
    get:
      operationId: listChores
      tags: [chores]
      summary: List all chores in the current household
      responses:
        "200":
          description: Chore list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chore"
    post:
      operationId: createChore
      tags: [chores]
      summary: Create a new chore (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreInput"
      responses:
        "201":
          description: Chore created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chores/{choreId}:
    patch:
      operationId: updateChore
      tags: [chores]
      summary: Update a chore (admin only)
      parameters:
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChoreUpdate"
      responses:
        "200":
          description: Updated chore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteChore
      tags: [chores]
      summary: Delete a chore (admin only)
      parameters:
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chores/{choreId}/rotation:
    patch:
      operationId: updateChoreRotation
      tags: [chores]
      summary: Update the rotation order for a chore (admin only)
      parameters:
        - name: choreId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RotationInput"
      responses:
        "200":
          description: Updated chore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chore"
        "403":
          description: Admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /assignments/{assignmentId}/done:
    patch:
      operationId: markAssignmentDone
      tags: [assignments]
      summary: Mark a chore assignment as done
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "403":
          description: Not your assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Assignment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, name, email, householdId, isAdmin]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        householdId:
          type: ["integer", "null"]
        isAdmin:
          type: boolean

    # Households
...
    ChoreInput:
      type: object
      required: [name, frequency, rotationOrder]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, custom]
        customDays:
          type: ["string", "null"]
          description: Comma-separated weekday numbers for custom frequency (e.g. "1,3,5")
        rotationOrder:
          type: array
          items:
            type: integer
          description: Ordered list of user IDs
    ChoreUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, custom]
        customDays:
          type: ["string", "null"]
    RotationInput:
      type: object
      required: [rotationOrder]
      properties:
        rotationOrder:
          type: array
          items:
            type: integer
    Chore:
      type: object
      required: [id, householdId, name, frequency, rotationOrder]
      properties:
        id:
          type: integer
        householdId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        frequency:
          type: string
        customDays:
          type: ["string", "null"]
        rotationOrder:
          type: array
          items:
            type: integer

    # Assignments
    Assignment:
      type: object
      required: [id, choreId, choreName, userId, userName, date, status]
      properties:
        id:
          type: integer
        choreId:
          type: integer
        choreName:
          type: string
        userId:
          type: integer
        userName:
          type: string
        date:
          type: string
        status:
          type: string
          enum: [pending, done]
        confirmedAt:
          type: ["string", "null"]

    # Expenses
...