# lib/api-zod/src/generated/types/chore.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Roomie API - Chore rotation and expense splitting for shared households
 * OpenAPI spec version: 0.1.0
 */

export interface Chore {
  id: number;
  householdId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  frequency: string;
  /** @nullable */
  customDays?: string | null;
  rotationOrder: number[];
}