# lib/api-zod/src/generated/types/choreInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Roomie API - Chore rotation and expense splitting for shared households
 * OpenAPI spec version: 0.1.0
 */
import type { ChoreInputFrequency } from './choreInputFrequency';

export interface ChoreInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  frequency: ChoreInputFrequency;
  /**
     * Comma-separated weekday numbers for custom frequency (e.g. "1,3,5")
     * @nullable
     */
  customDays?: string | null;
  /** Ordered list of user IDs */
  rotationOrder: number[];
}